backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-search-documents
openai
tiktoken
pypdf